requirements.txt
Django==1.9
psycopg2
django-html5
